# public/api-doc.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Makkelijke Markt API",
        "version": "1.1.0"
    },
    "paths": {
...
        "/api/1.1.0/login/basicId/": {
            "post": {
                "tags": [
                    "Login"
                ],
                "summary": "Genereert een nieuw token op accountId + password",
                "operationId": "LoginPostByAccountId",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "accountId",
                                    "password"
                                ],
                                "properties": {
                                    "accountId": {
                                        "description": "Account ID"
                                    },
                                    "password": {
                                        "example": "string"
                                    },
                                    "deviceUuid": {
                                        "description": "UUID van het gebruikte device",
                                        "example": "string"
                                    },
                                    "clientApp": {
                                        "description": "appliciatie type",
                                        "example": "string"
                                    },
                                    "clientVersion": {
                                        "description": "Versie van de client",
                                        "example": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Token"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "description": "",
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "description": "",
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "description": "",
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "423": {
                        "description": "Locked",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "description": "",
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "api_key": []
                    }
                ]
            }
        },
        "/api/1.1.0/login/basicUsername/": {
            "post": {
                "tags": [
                    "Login"
                ],
                "summary": "Genereert een nieuw token op username + password",
                "operationId": "LoginPostByUsername",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "username",
                                    "password"
                                ],
                                "properties": {
                                    "username": {
                                        "description": "Account Username"
                                    },
                                    "password": {
                                        "example": "string"
                                    },
                                    "deviceUuid": {
                                        "description": "UUID van het gebruikte device",
                                        "example": "string"
                                    },
                                    "clientApp": {
                                        "description": "appliciatie type",
                                        "example": "string"
                                    },
                                    "clientVersion": {
                                        "description": "Versie van de client",
                                        "example": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Token"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "description": "",
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "description": "",
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "description": "",
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "423": {
                        "description": "Locked",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "description": "",
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "api_key": []
                    }
                ]
            }
        },
        "/api/1.1.0/login/whoami/": {
            "get": {
                "tags": [
                    "Login"
                ],
                "summary": "Geeft eigen account informatie",
                "operationId": "LoginGetWhoAmI",
                "responses": {
                    "200": {
                        "description": "",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "items": {
                                        "$ref": "#/components/schemas/Account"
                                    }
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "api_key": [],
                        "bearer": []
                    }
                ]
            }
        },
        "/api/1.1.0/login/roles/": {
            "get": {
                "tags": [
                    "Login"
                ],
                "summary": "Geeft lijst mogelijke rolen",
                "operationId": "LoginGetAllRoles",
                "responses": {
                    "200": {
                        "description": "",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "ROLE_USER": {
                                            "example": "Gebruiker"
                                        },
                                        "ROLE_ADMIN": {
                                            "example": "Beheerder"
                                        },
                                        "ROLE_SENIOR": {
                                            "example": "Senior gebruiker"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "api_key": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
            "Account": {
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": "14"
                    },
                    "naam": {
                        "type": "string"
                    },
                    "email": {
                        "type": "string"
                    },
                    "roles": {
                        "type": "array",
                        "items": {
                            "type": "string"
                        },
                        "example": [
                            "ROLE_USER"
                        ]
                    },
                    "username": {
                        "type": "string"
                    },
                    "locked": {
                        "type": "boolean",
                        "example": false
                    },
                    "active": {
                        "type": "boolean"
                    }
                },
                "type": "object"
            },
...
            "Token": {
                "properties": {
                    "uuid": {
                        "type": "string",
                        "example": "aaa9d3dd-bbb6-4d99-bca1-bfdb1aaa5a35"
                    },
                    "creationDate": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "lifeTime": {
                        "type": "integer"
                    },
                    "timeLeft": {
                        "description": "in seconds",
                        "type": "integer"
                    },
                    "deviceUuid": {},
                    "clientApp": {},
                    "clientVersion": {},
                    "account": {
                        "$ref": "#/components/schemas/Account"
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "api_key": {
                "type": "apiKey",
                "name": "MmAppKey",
                "in": "header"
            },
            "bearer": {
                "type": "apiKey",
                "name": "Authorization",
                "in": "header"
            }
        }
    },
...
}